Redirects can be reformatted.
The osh AST now only stores file descriptors that appear explicitly in
the source code.  At runtime we fill in defaults (this can be done at
compile time later.)

Lots of planning and notes for other conversion.

Add a test for scopes.  We need to do semantic analysis to properly
convert shell scripts!