Big rewrite/simplification of simple command evaluation and redirects.
This fixes the issue argv[0] was evaluated in the parent process instead
of the child.  This is how bash/dash/mksh behave (but not zsh).

We had all these unnecessary Thunk types which are now mostly gone. The
old code was embarassingly complicated and incorrect: a case of
optimization before correctness.

Instead of an OOP representation of redirects, we use functional one
expressed with ASDL (redirect in runtime.asdl).  This style is
consistent with the rest of the code.

Also:

- Implement redirects at function definition site
- Properly implement waiting for the here doc process.

Misc:
- Fix bug in invoking gawk and add comment.
- Adjust test failures