Got the BuiltinFlags parser working.
This style of flags parsing allows for:

cut -d,

i.e. no space between the flag and arg.

It also allows the very odd:

    read -rp '>'

Where -r is a flag with arity and -p is a flag with arity 1 that takes
'>' as an argument.

I think the unit tests cover almost all cases.