Redesign the Mem interface.
This allows correct implementation of 'readonly', 'unset', 'export',
'export -n', array LValues, including mutation with with ++ and --, etc.

The core of it is the new mem.SetVar() interface.  It has parameters for
var_flags and the lookup_mode (scope).

There are helper functions in state.py which use the public interface of
Mem.

We use dynamic scope now, but it should be easy to change to
local/global scope.

Notes:

- unset readonly variable has stricter behavior than bash because it's a
  programming error.

Also:

- Notes on array representation.
- Added exhaustive tests in state_test.py.
- Fix builtin function naming scheme