Change the strategy for integrating ./configure output and the Makefile.
You can now build and test the tarballs.  The tarballs are tested by
unpacking them, running ./configure, and make.

Instead of having different lists of C dependencies, just remove
readline.c in the compile action if HAVE_READLINE is not true, and add
an #idef HAVE_READLINE in the generated module_init.c.

The original strategy (using static-c-modules.txt) is problematic
because the build will depend on Python.

Details:
- change xargs invocation so it will fail on nonexistent input