Draft of the OSH quick reference and associated tool.
There is a *-toc.txt format and a *-pages.txt format.  The idea is to
write one doc that is both the online / in-shell help and the quick
reference "manual", rather than having two like bash.

Actually bash seems to have THREE pieces of documentation -- the 'man'
page covers 'times' but that is different than in-shell help!

dash, mksh, and zsh don't seem to have any in-shell help.

Also:

- Build the release binary, and test it out with spec tests.  Make note
  of a non-deterministic crash I saw.
- Get rid of old comments in core/builtin.py that pertain to flag
  parsing.  That's now handled in core/args.py.