Fixes to the test / [ builtin.
I learned that the [ language isn't really specified by a grammar, as [[
is.  Instead we handle it like bash does: by splitting it into 6 cases:
expressions of length 0, 1, 2, 3, 4, and more.

The bash source implies that POSIX specifies this.  And it explains why
-a and -o are deprecated.

I'm not supporting -a as a unary operator, since it's an alias for -e.

I also added a couple test cases to dbracket.test.sh, and fixed some
syntax error handling bugs in [[.

NOTE: There's still a bug where [ abc = 'a*' ] should not support
globbing.

Addresses issue #19 and issue #26.