Wow, greatly simplified here doc parsing code!
It should have sped it up too, but I still have to measure it.

The post-order traversal was the wrong idea.  All we need is simple
deque().  (It's probably a linked list in C.)

Also, fix here strings/here words to evaluate to a string with an
additional newline.  gold/complex-here-docs.sh now passes!