Introduce the concept of asdl.const.NO_INTEGER.
It is the default value for unspecified ASDL integers.  See the comment
in const.py for the rationale.

Change span_id and line_id to be 0-based, with const.NO_INTEGER as the
uninitialized/unknown value.

All unit tests and spec tests pass.

Also:

benchmarks/oheap.sh: Try to encode everything in benchmarks/testdata.

This revealed that we should properly encode Array<Str>, which is now
done.  However there are still some lingering negative numbers.