Generate Python source code from an ASDL schema.
This is in contrast to using metaprogramming at runtime.

OSH works with this new code, and it's a lot faster!  It's not enabled
yet because a few unit tests still fail, due to lack of __eq__.  But the
shell runs and passes spec tests.

Also:

- Decode the output of everything we encoded in the oheap benchmark.

Document an idea to both measure decoding speed and collect data about
Array/Str lengths.  This could help with record sizing, inlining, and
other decisions that affect locality.

- In ASDL constructors, Raise TypeError instead of AssertionError.  This
  is more consistent with Python itself.