Fix bug parsing test expression (found in tcc's and QEMU's configure)
The expression is of the form:

'test -z x -o z = z'

It only happened in expression longer than 4 tokens to trigger the
osh/bool_parse.py rather than the heuristics.

The bug was that (BoolBinary_Equal '=') was assigned the wrong kind,
because it is setup after all the other id / kind pairs.

Added a unit test.  All spec tests pass.

Also:
- Polish the error message when there is a bad 'test' expression
- Show the files that a configure script changed in
  benchmarks/runtime.sh.