Fix bugs in pushd / popd / dirs.
The stack was printed in the reverse order.  Until the last commit, the
spec tests didn't expose this.

Printing os.getcwd() in 'dirs' isn't correct.  The correct semantics are
that the dir_stack is initialized to the current directory, and it is
reset upon the 'cd' builtin.

popd fails if the stack is of length 1.

Tests all pass now.