Make the behavior of top-level return/break/continue follow dash/mksh.
Bash's behavior is inconsistent.  If you can return from a source'd
script, you should be able to return from the main script.

This behavior is OK for running Aboriginal Linux.  But the old behavior
of disallowing return at the top level was not OK, and it didn't match
that of any shell.

Also add set -o strict-control-flow -- the first STRICT option.  It
disallows 'continue' and 'break' at the top level, but allows 'return'.