Figuring out call stack introspection in bash via spec tests.
It appears the variables that collectively store the call stack don't
line up!  ${BASH_SOURCE[i]} and ${BASH_LINENO[i]} are out of whack!

But the current position is ${BASH_SOURCE[0]} and $LINENO.  Sigh.

This was good to test out, but I don't think Oil will compatible with
bash (and hence debuggers like bashdb).

Also:
- Test the values of $$ and $BASHPID in command subs too.