Temporary env bindings are not visible during function calls.
That is, 'FOO=bar myfunc' now works.  This bug was tickled by Aboriginal
Linux.

Prior to this change, I constructed an 'environ' dict from the env
bindings, and only passed that to external processes.

Now there is a read-only _StackFrame() on the variable stack.

It can be (1) read from and (2) set with the new scope_e.TempEnv mode.
LocalOnly and Dynamic scope skip it.  The 'read' builtin uses LocalOnly
so the binding persists.

In the spec tests: a very good exploration of temporary env semantics.
dash, bash, and mksh all behave differently!

No spec tests regressions.

Also:

- Basic xtrace tracing for Assignment.  In OSH, assignments are
  different than commands.