Make travis perform developer build, and {unit,spec} tests.
* Add the dependencies from `build/dev.sh ubuntu-deps` and `test/spec.sh
install-shells` as packages for container (according to travis
documentation).
* Run necessary script `test/spec.sh link-busy-box-ash` before
installation.
* Create `_tmp` directory that is used by tests.

Although it's possible to install the dependencies through the script
`build/dev.sh ubuntu-deps`, that'd imply that the integration has to be done
using a VM instead of a container (Because the script uses sudo). I'm gonna
try to make the integration using a container, which is considerably faster.
The downside is that dependencies have to be specified both in the script and
in the travis file. If it fails, then it's as easy as substituting the
`addons.apt.packages` lines for:

sudo: required

before_install:
    - build/dev.sh ubuntu-deps