Makefile: Put either of bytecode-{cpython,opy}.zip in the app bundle.
Also, move most rules to portable-rules.mk, which is intentionally NOT
included in the release tarball.  Now we are less confused about a
developer build (out of the repo) vs. an end-user build (out of the
release tarball).

This was motivated by a bug I wasn't able to figure out: For some
reason GNU Make tried to build oil-version.txt because it's a
prerequisite of _build/oil/bytecode-cpython.zip, even though the latter
exists, and it didn't do this in oil-0.4.0.

Also:

- _build/osh_help.py moved to _devbuild/gen/osh_help.py, so
  _build/osh-quick-ref/ should also move to _devbuild/osh-quick-ref.