Remove obsolete util_opy.py.
We are now sharing core/util.py.

Surprisingly, this introduced a change in the compiled output of
_regtest/src/util_opy.py!  The compilers source code leaks through to
the thing being compiled!

I think this must because of the way consts are handled.  When you are
compiling a const, it might be looked up in global interpreter state.
If the same constant is not in the interpreter because util_opy.py is no
longer imported, then the results can be different.

Haven't fixed this diff yet:

$ ./regtest.sh verify-golden
--- _regtest/dis-md5.golden.txt 2018-03-17 20:12:45.245834622 -0700
+++ -   2018-03-17 23:11:28.941012975 -0700
@@ -34,7 +34,7 @@
    823 5a440f929760bbd1f36d3bfa4b3779f8 _tmp/regtest/core/word_eval_test.pyc
    834 4a3120879710a4fd9f9d6039e4dade55 _tmp/regtest/opy/compiler2/consts.pyc
    845 0eea196e032e82f3ddc932265ce3b640 _tmp/regtest/asdl/arith_ast.pyc
-   909 b82e6bdb2652220742231fbe2bd14221 _tmp/regtest/opy/util_opy.pyc
+   899 0dd9d5f5018f43b257a438ad9680d1c1 _tmp/regtest/opy/util_opy.pyc
   1010 338f3f9966140f857e952f610a3918db _tmp/regtest/spec/bin/read_from_fd.pyc