Fix segfault under gcc 8.
CPython had undefined behavior for a long time, and gcc 8 turned this
into a segfault.

This commit description explains it well:

https://github.com/python/cpython/commit/e348c8d154cf6342c79d627ebfe89dfe9de23817

There was a 'long double' to ensure a larger alignment (which apparently
mattered for some platforms, at some time in the past).  But Python's
memory allocator only guarantees 8-byte alignment, which caused the
undefined behavior.

Other background:

https://www.mail-archive.com/gcc-bugs@gcc.gnu.org/msg570352.htmlo

https://mail.python.org/pipermail/python-dev/2018-January/152000.html

Fixes issue #107.