Initial implementation of alias.
This isn't totally correct since multiple alias expansion aren't
handled.  But we go from 22 test cases failing to 6!

The alias expansion is detected in CommandParser.ParseCommand(), and
then we push a buffer onto the Lexer at that point.  This seems to be
the correct place to do it.

Addresses issue #160.