A different approach to implementing alias.
The previous approach of PushAliasBuffer() only worked for a single
alias expansion.  Iteration was a problem.

Now we hook into ParseSimpleCommand and expand aliases iteratively.
We join the expanded aliases with the rest of the *text* from the
SimpleCommand, and then pass that into ParseCommand().

I made some comments about the cases that this doesn't handle.  And also
some notes about how dash and bash do it -- with global variables strewn
throughout the parser.

(Although I wonder if we could do that with self._Next() ? )

- Add more alias test cases
- Factor some free functions out of CommandParser.

All unit, spec, and gold tests pass.