Add regex escaping to [[ foo =~ $regex ]].
Right now there is a hack to only escape {}, which was determined
empirically.  It gets us past a certain expression in the
'bash_completion' main script.

Apparently () and | don't need to be escaped, and neither to * + and \.

We might want to do something more principled later.  But I added a
couple test cases.