[printf] Crude first version of printf builtin.
Passes one spec test!

And good enough to run bash_completion, and now without the message
  -vprintf: warning: ignoring excess arguments, starting with ‘quoted’
we were seeing, which is from /usr/bin/printf not understanding `-v`.

This shadows /usr/bin/printf, which has more functionality but also
differs from the shell builtin.  So some scripts that worked may now
fail, in which case the fix is to add the missing functionality here.