[completion] Try to properly quote filename completions.
This exposed an issue with using the default READLINE_DELIMS.  I
sketched out a plan in demo/pyreadline.py to use no delimiters.

It appears we return FULL LINES as candidatse, and use a custom
completion hook to display the current word -- where words are
determined by OUR OWN parser and not readline.

- spec/builtin-completion: Another test case.  It changed between 4.3
  and 4.4.
- demo/tty.sh: learned how to use ttyrec and script.  Learned a cool
  named pipe trick with script!  Piped it into 'vtparse'.