[completion] 'compgen -X pattern' now combines with other flags.
The issue is that non-interactive bash shells don't respect -X, but we
only care about interactive shells.

Had to work around this in the spec tests.

Also: add spec test cases for details of compgen:

- -W respecting IFS
- filenames with funny characters

As much as possible, we want to pull all quoting an escaping out of
bash-completion and put it in OSH.  The OSH binary knows the language!