[build] Detect objcopy and not dsymutil.
We need objcopy to put symbols for the OPTIMIZED binary in a separate
file, and then add a "debuglink" field in the ELF.

If we don't have objcopy, e.g. with Clang on OS X, we just strip it and
lose the symbols.  It doesn't appear to have the "debuglink"
functionality?

Should fix issue #237.