[interactive] Ctrl-C now prints ^C and cancels a command.
This is what other shells do.  Prior to this change, Ctrl-C was always
ignored.

Addresses issue #36.

Also, a significant refactoring: Put the prompt/display cleanup at the
END of the interactive loop.  It was at the beginning for silly reasons
of code structure, but the duplicate Reset() was redundant and caused
problems on one pull request.