[prompt] Prettify expansion of \W with ~ where applicable.
The logic is shared with the 'dirs' builtin and moved into core/ui.py.

This led to a cleanup of $HOME.  OSH no longer sets $HOME on startup.  I
mistakenly thought that bash always did this, but it only does it for
login shells (in which case it's exported).

Fixes issue #246.