Fix crash in 'builtin command ls'.
I had put 'command' in an odd place in the Executor because I didn't
want to thread (fork_external, span_id) through so many functions.

But that caused a bug in the odd case of 'builtin command ls'.  Now
we're doing it the obvious way.  Everything fits together correctly.

In the future, we want to:

- have argv and span_id travel together.  The evaluated argv array needs
  location info.
- fork_external needs tests (issue #254).  Maybe make it a flag on the
  Executor instead of a parameter to so many methods.