Export PWD and OLDPWD.
That is, the variables are not only set within the shell, but exported
to child processes.  This seems to be an undocumented feature of bash
and dash?

Found with

env -i -- $sh -c env

Tricky issue: it's hidden in spec tests with bin/osh, because bin/osh
is a shell wrapper run with dash (/bin/sh)!  It only showed up in the
tests against the release binary, where we use a symlink.