[benchmarks/osh-runtime] Don't pass --parser-mem-dump.
As of the last commit, it's only valid on -n.  This is a result of last
year's execution model change, due to alias.

- Fixed a bug with OSH_OVM getting $PWD prepended to it twice.
- Get rid of old oheap stuff.
- Add better error message to benchmarks/time.py

This still has to be tested end-to-end, because unfortunately the
reporting pipeline assumes you have two machines.

TODO: fix that and make a way to test the reports end-to-end on a single
machine.