[plugins] PS1 and PS4 can no longer modify $? or ${PIPESTATUS[@]}.
Fixes issue #268.

state.Mem() now has a stack of status values.  word_ev.EvalForPlugin()
pushes a frame onto it before evaluation, and pops it afterward.

So the code inside sees the correct values, but the state of the shell
is unaffected.  This is similar to what bash does, but cleaner IMO.

demo/plugins-and-shell-state.sh has some manual test cases for this.

(TODO: Fix unrelated PS4 bug and add a spec test.)