[word-language] Treat ${var%[} as bash and busybox ash do.
If var='[foo]', then we will get 'foo]'.

This edge case is pretty confusing (even by shell standards.)

For some reason, dash, mksh, and zsh disagree.  They must treat [ as an
invalid glob and leave it as '[foo]'?  In other words, they require
${var%'['}.

But bash and busybox ash agree, and the code seems to make sense -- i.e.
we perform glob unescaping if the arg word doesn't look like a glob,
because we already glob-escaped it.  LooksLikeGlob() is necessarily a
heuristic.

Fixes issue #290.