[types] Add typechecking for osh/builtin_bracket.py
Also add --follow-imports=silent to typecheck-more-oil (see comment
there), and remove the assert-one-error hack which is no longer
needed.

Had to add an ugly annotation in bool_parse.py, but I think the
solution might be to refactor the code itself a bit.  Then again, the
code may just be reflecting the ugly reality of parsing bash.