Add annotations and enable typechecking for osh/word_compile.py and osh/builtin_bracket.py (#542)
* [types] add some files that already pass mypy to typecheck-more-oil

* [types] add type annotations to osh/word_compile.py

* [types] move `typecheck-more-oil` filenames into their own manifest

* [types] add command `types/run.sh need-typechecking` to show files not yet covered by mypy

* [types] Add typechecking for osh/builtin_bracket.py

Also add --follow-imports=silent to typecheck-more-oil (see comment
there), and remove the assert-one-error hack which is no longer
needed.

Had to add an ugly annotation in bool_parse.py, but I think the
solution might be to refactor the code itself a bit.  Then again, the
code may just be reflecting the ugly reality of parsing bash.