[performance] Avoid allocating Token.val for 5 kinds of tokens
This reduces the number of allocations / bytes measured by dumb_alloc.

Though that underestimates it, because we also use malloc() in the Str
implementation.

You can see LookupKind goes up in the uftrace profile, and memcpy goes
down.

I think it probably makes performance a bit better, but the benchmark is
noisy and it's within the noise.  But using less memory is better.