[osh-language] Implement the nameref feature (declare -n)
Addresses issue #600.

There are now 22 spec tests in spec/nameref.test.sh.  There's one Oil
failure related to the combination of -n and ${!ref}.  Oil and bash
disagree, but I don't understand bash's behavior and it's hard to copy.

There is also 'shopt -s strict_nameref' which makes it behave more like
mksh.

This feature is full of pitfalls as usual, but there's a sane subset
that's shared between bash and mksh.