[osh-language] Change the 'unset' builtin to delete the cell.
We don't set the cell to Undef.

Changed assertions in 'spec/assign'.

We no longer match dash/zsh in most cases.

- Case 24 we have unique behavior, but we always did (/)
- We match bash with one case
- We match mksh with two cases

Addresses issue #706.

This is sort of an experiment.  Let's see if ble.sh runs.  Either way
the semantics are simple, and this appears to be closer to what it
wants.