[benchmarks/alloclog] Estimate overhead of garbage collection.
Bottom line: Oil is written in a very allocation heavy style, so the
performance of both parsing and runtime depends a lot on the allocator.

Mark and Sweep is looking unappealing.  Not only does it take us back to
the system allocator (if you want a non-moving GC), there's Too much
object overhead.

Research to do on copying GC: I want to see if we can solve the "double
indirection" problem with something like v8::Local.