[oil-language] Detect duplicate declarations as syntax errors.
This includes proc params.

Wasn't too hard!

We should also be able to test set/setlocal statically.  But I
think setvar/setglobal don't know what globals there are because of
'source' and environment vars.  (shopt -u copy_env in Oil?)

This fixes the Julia example.

But we also have a 'const' issue.  I think that is OK for now.