[opy/regest] Update golden checksums
I think they must have all changed because I'm now on Ubuntu Bionic,
which has Python 2.7.17?

Honestly thing whole thing is old and we should get rid of it.  I don't
think it has caught any bugs?

I suppose it's sort of an arbitrary restriction on Oil's codebase
though.  But mycpp and static typing are even stricter restrictions.

On the other hand, we ARE still using OPy for the tarball build, so I
guess we should keep maintaining this.