[oil-language] Implement shopt parse_triple_quote
This allows the following in COMMAND mode:

    """
    '''
    r''' (when parse_raw_string is also on)
    $'''

We use a little hack where the token for the third quote has its 'id'
field mutated into Id.Left_TSingleQuote or Id.Left_DollarTSingleQuote.

This means that the left token is different in command mode or
expression mode.

Update the tests.

Still need to properly strip leading whitespace in both command mode and
expression mode.