[cpp] Add unit test for pyos::ReadByte()
This uncovered the fact that the macro hackery for posix::O_RDONLY and
errno_::EINTR wasn't working.  And it was causing problems in the tests.

So I removed all of it, and we just use unqualified EINTR like in
C-style code.  We're doing this for WUNTRACED too, so now it's
consistent.