[fix] Fix lookahead for ( in WordParser
This was introduced in the commit that started using ByteLookAhead()
instead of LookPastSpace().  We need the new LookAheadOne() method.

In regex mode, ( can be a literal!  So we have to look for Id.Op_LParen
and not '(' literally.

TODO: Look at the other ByteLookAhead() calls to see if they have this
problem.

Test in spec/regex now passes.