[errors] Add errfmt.PrintMessage(), and make errors more consistent.
Sometimes we want to print location info without showing the code.  It's
too verbose in the case of builtins.  It makes sense when there is a
usage error, but I think we should leave it off for runtime errors.

Makes the error messages in #1044 less verbose.  dash and zsh don't
print ANY error, and just return non-zero.  bash and mksh print an
error.