[build] Start porting oil-native build to Ninja.
This eliminates redundant work in the soil/cpp CI task.  You can just
demand the binary without worrying about whether you already built it.

It should also reduce work in the release process, although we have to
deal with the tarball build there.

We're following the pattern we used in mycpp's Ninja build:

- build/native_graph.py generates a build.ninja file
- build/mycpp.sh contains build steps, althuogh these should be moved to
  a different file
- build/native.sh is a wrapper