[interactive] Ignore SIGWINCH in the shell.
The line editor is the only thing that needs to know the width, so I
think we should let it handle the problem.

Unfortunately GNU readline is not what we want, but that's a separate
issue.

This means the core/comp_ui.py now doesn't adjust when the window is
resized.  But that dosn't make the shell unusuable.  We should revisit
it later.

Apparently fish and zsh do something different anyway -- they REDRAW on
SIGWINCH.  They don't merely update the width.

- Added a test in test/interactive.py.
- Addresses #1067