[comp_ui] Keep track of screen width in GNU readline.
OSH now queries this state, instead of trying to keep track of it
ourselves.

So OSH behaves like bash and other shells here:

- If trap 'echo X' SIGWINCH, then SIGWINCH will interrupt the wait
  builtin.
- Otherwise it is ignored.

We take care to only enable the SIGWINCH handler inside call_readline(),
when shell code isn't running.

When the TAB completion callback is invoked, we do the reverse: we
disable our SIGWINCH tracking.

So we can lose some resizes, but this is what other shells do.  TODO:
Test this out a bit more.

Addresses #1067.