[signal-handling] Ignore untrapped SIGWINCH on wait.
This fixes #1067 for real this time!

I use a slightly odd technique of setting SignalState::last_sig_num to
the sentinel UNTRAPPED_SIGWINCH.  We still need to redesign signal
handling for the C++ translation.

Add a failing test for the analogous 'wait -n' bug, which still needs to
be fixed.  It needs to run in a loop until a state is reached.