[doctools] Add spell checker
We avoid all the false positives and adding words like "HTML" to the
dictionary.

The first pass finds words that appear exactly once in the whole
corpus, and which are not in /usr/share/dict/words.

This was enough to find many typos!

We use 'lynx -dump' to convert HTML to text, and then a custom word
splitter.  Markdown would have extra URL tags and so forth.