[doc] Polish OSH help and Oil help
This system may not be scalable.

I think the tight indices are useful for me: {osh,oil}-help-topics.  But
there needs to be some other method of organization for {osh,oil}-help.

And I don't think they will underlie the help builtin.  The help builtin
might show URLs or something?  I think it will be like the fish shell.

I think we want to emphasize introspection and self-documentation.