[oil-language] Fix bug where d->data conflicted with 'data' keyword
Now we have a 'tea_keywords' boolean in the expression parser.

I think a new lexer mode might also work?  We can try that later.

We sacrificated one test case in tea/testdata/func.tea -- for nested
first class functions.  Not sure why, but it's not important now, as Tea
is far in the future.