Revert "[refactor] Remove wrapper in oil_lang.objects"
This reverts commit 08782be6e7a0fb521aa7f83babe721cab1b55d99.

There were still bugs; it would be like playing whack-a-mole.  We need
to do #636 the right way.

The core issue is that value.MaybeStrArray() and value.Obj(mylist) have
to be reconciled thoroughly.

And ditto with value.AssocArray() and value.Obj(mydict).  There's no
good way to do that with only localized changes.  It affects all
operations.