[oil-language] Multiple for loop variables are separated by commas
Part of #1135.  This should finish it.

We technically allow space-separated variables, but it's not preferred.

This style is more familiar to Python and Go users.  It also creates a
syntactic distinction between loop variables and words to iterate over:

    for i item in a b {  # these look too similar
      echo
    }

Better:

    for i, item in a b {
      echo
    }

Later we could also add

    for i, item in %(a b) {
      echo
    }

But it's probably not necessary.