[oil-language] First cut of Config Dialect, called Hay :)
Sketched it in doc/hay.md.

Still much to do but this gives us the core semantics / evaluation model.

It builts a tree with two types of nodes: attribute / data nodes, and
shell / code nodes.  Declarative and imperative interleaved.

Unfortunately it's not translated to C++ because of the Dict[str, Any]
and JSON issue.