[oil-language] Remove placeholder for parse_amp
This means that 'sleep 1 & wait' works again, following the principle of
not breaking the top level.  The 'fork' builtin is still idiomatic.

This means we're not doing #832.  The idea in #1059 to recast
redirections as imperative is more promising, and won't conflict with
the syntax.

Now that we have

    fopen > out.txt {
      echo 1
      echo 2
    }

It makes sense to have something like:

    redirect {
      push '>' 1 out.txt
      push '&' 2 1        # because >& and <& are actually the same!

      ls /tmp
      echo hi
    }

As a flexible alternative to:

    { ls /tmp; echo hi; } >out.txt 2>&1