[oil-language] Only allow bare assignment inside Caps { } blocks.
It's no longer an Oil / Hay distinction.  This is good because it
reduces parser parameters/state!

It's also good because we disallow 'x = 42' in most places in Oil.  We
don't parse it as the command 'x' with arg '='.

And move shopt --set parse_equals to the oil:upgrade group, because bare
'=' shouldn't be common.

Part of #951.  Related to previous parse_equals changes in this release.