[build] Fix macOS build from source failing at make step
macOS release build of oil-0.10.1 throws an error on the `make` step.

This is with oil-0.10.1.tar.gz from the website; have not checked if the
brew version suffers from the same issue.

After running `make`, from Python-2.7.13/Python/_warnings.c:
> implicitly declaring library function 'tolower' with type 'int (int)'

with a helpful note to:
> include the header <ctype.h> or explicitly provide a declaration for 'tolower'

This appears to be the same error mentioned in #978 in the
“So I attempted to install from source” section.

There’s a bunch of warnings too and then at the strip substep there’s a
final error because:
> can't open file: _build/oil/ovm-opt (No such file or directory)

After nuking everything I included <ctype.h> then started the build steps
again to great success.