[mycpp] Work on examples/parse
It translates, compiles, and passes test!  I fixed the ASDL build issue
by generating manifests with build/app-deps.sh.

And now it crashes in the benchmark with an assertion, which is the bug
we want to fix.  ASDL types don't have field masks!

We need to address #1164 and #1176 to make this pass.