[cpp] New string creation API that's common between leaky and GC.
And use the new API in both modes in the QSN binding.  This simplifies
the code and migration strategy.

Still TODO: make a gc_migration_test.cc.

Summary:

When you know the length

    Str* BlankStr(int len)

For snprintf() style APIs:

    Str* OverAllocatedStr(int len) and s->SetObjLenFromStrLen()

For interfacing with C code:

    Str* CopyStr(const char* s);
    Str* CopyStr(const char* s, int len);

This API is tested from both mylib_test.cc and gc_heap_test.cc.