[fix] Fix nullptr deref in List::reserve()
Caught by ubsan:

/home/c/dev/oil/mycpp/gc_heap.h:972:47: runtime error: member access within
null pointer of type 'Slab<int>'

Previously, List::reserve() was copying from the old slab to the new one
whenever the *newly updated* capacity was greater than zero. If the
capacity was previously zero, slab_ might be nullptr.

Now it copies only when the list is non-empty.