Fix list_contains with specializations (#1214)
* Change list_contains() to be safer at compile time

The spec test `alias -r 3` was experiencing infinite recursion due to
list_contains() never returning true.

The bug was that the most generic template specification compared
elements using `operator==`, which on pointers defaults to pointer
equality.

The offending line: `if (haystack->index_(i) == needle) {`

What was happening is the type `Tuple2<Str*, int>*` was getting
substituted in on that generic template because there was no
specialization that prevented it (such as in the case of `Str*`).

I experimented with several different methods of systemically fixing
this issue.

I first tried to overload the Tuple `operator==` in several different
ways, none of which would prevent the bug.

I realized that the issue is not necessarily with the Tuple type, but
with the `list_contains` definition itself.  The way it was written
was 'generic' in the sense that it works if everything was implemented
correctly (in this case we needed an additional list_contains def such
as the following):

bool list_contains(List<Tuple2<Str*, int>> *haystack, Tuple2<Str*, int>  *needle);

But it fails to report when we actually _need_ to implement additional
list_contains functions.

After some deliberation I settled on simplifying the definition of
list_contains to be totally generic, and check for equality using
hand-rolled `are_equal()` functions.  This has an obvious drawback that
we need to implement `are_equal()` for primitive types, but has the
significant strength that it breaks at compile time if we do not have a
comparator defined for an instantiated type.

Before settling on this method I attempted several times to overload
`operator==` in clever ways such that primitive types would work but
unimplemented types would break.  I was unable to find an effective
solution, and I believe that using that method is actually infeasible
for preventing this bug.  To use the `operator==` method you need to be
able to tell the compiler to complain if it's asked to compare two
pointers, which I'm not sure is actually possible.  C++ templates are
supposedly turing-complete, so it should be _possible_, but I have no
idea how one might do that.