[soil] Add debugging for podman disk space issue.
It's using 12 GB of space for a 991 MB image!  (Also reported by 'podman
images' as 1.56 GB for some reason.)

This is why we want containers to just be "plain old data".  (It would
be nice to try bubblewrap.)

Disable the 'dummy' task for now, since 'cpp-coverage' takes the entire
sourcehut VM because of this issue!