[translation] Translate exec / umask / trap builtins
This is a first pass -- some functions are stubbed out as
NotImplemented(), and we also have some 'if mylib.PYTHON' guards.

Now we can run Python's configure with bin/osh_eval.py!

Part of #1165.

Also add builtin_process to the build/app-deps filters, since it's not
ready to be translated.