[mycpp/runtime] Start showing the f(g(), h()) problem
With Str::slice().  Demonstrate it the old way (local var rooting), and
the new way (return value rooting).

This led to a change in the API: We now have RootSet::EndProcess(),
which calls RootSet::Clear() and then MarkSweepHeap::Sweep().

We're still hidden behind #ifdef, and nothing passes yet.

But I think it can pass with this implementation.  StrFromC() still
needs return value rooting.