[osh] avoid try/except/else in Umask
mycpp can't generate matching c++ for these. Since the part of mycpp
that would reject code like this is currently commented out (buggy
perhaps?), the control flow of the generated code doesn't match the
source, causing us to hit an unreachable code path in the umask test.
This was causing a spurious test failure.