[core] move trap-related state into SignalState
This is the first commit in a series that will refactor the signal and
trap handling code to make it safely shareable between ptyhon and C++.

The first step in this process is to remove the illusion that the
various users of the trap-related state each own it. We can make it more
obvious that they're really borrowing global state by renaming the
member(s) they all use to access it.