[core] replace _TrapHandler and SigwinchHandler
These types are nearly identical when it comes to signal handling. This
commit replaces their use as signal handlers with a generic signal
handler that simply enqueues signal numbers as they fire. The rest of
the handler logic (i.e. enqueuing AST nodes for execution in the main
loop) is moved to `SignalState.TakeRunList()`. This design will make it
easier to write the C++ signal handling code that is obviously
signal-safe. It will also allow us to make the bulk of the trap handling
code into translatable python (both coming in a future commit).