[core,osh] Make SignalState translatable (#1385)
* [core/pyos] remove defunct field from SignalState

* [core] replace _TrapHandler and SigwinchHandler

These types are nearly identical when it comes to signal handling. This
commit replaces their use as signal handlers with a generic signal
handler that simply enqueues signal numbers as they fire. The rest of
the handler logic (i.e. enqueuing AST nodes for execution in the main
loop) is moved to `SignalState.TakeRunList()`. This design will make it
easier to write the C++ signal handling code that is obviously
signal-safe. It will also allow us to make the bulk of the trap handling
code into translatable python (both coming in a future commit).

* [osh] remove _TrapHandler

This type is an empty husk now. Hook handlers can just be referenced by
their AST nodes directly.

* [core] move SignalState to builtin_trap

* [core/pyos] make SignalHandler a global

* [cpp] wrap signal handler state in a class

* [core] fix signal code style

* [core,osh]: merge SignalState and HookState