[frontend/consts_gen] Make string length explicit in lookup tables (#1394)
We currently use the `StrFromC(char *)` form for `Str` creation in the
generated character lookup tables for C++. This means '\0' gets
interpreted as the empty string instead of a string of length one with a
NULL byte in it.  Using the `StrFromC(char *, int)` form fixes this. For
characters other than '\0' this should make no difference.