[mycpp] Implement field reordering and Tag::Scanned for most classes
If a class doesn't use inheritance, then it's safe to reorder fields.

This was necessary because state::Mem and cmd_eval::CommandEvaluator
have more than 16 fields.

This gets rid of the annoying maskbit() compile warning.

Also fix bug with __init__ and HasVTable.  Add tests for it, and more
tests.

(There's still a crash with field masks we have to figure out and fix.)