[mycpp/demo] Prototype of new GC header.
It has a 24-bit object ID, which should allow us to get rid of the
std::unordered_set for the set of marked objects.

I learned that unions and bitfields don't pack as you'd hope.

Assert that ASDL compound sum types don't have more than 128 variants.
We have 7 bits in type_tag for this info.  Really the max is 38 variants
or so for syntax_asdl::command, so it would fit in 6 bits.

Mostly figured out big/little endian issues.

This is part of #1354.