[benchmarks/uftrace] Analyze string and alloc sizes with R
Hm it seems to show that most strings ARE small, less than 6 bytes.  But
most allocations are not strings!

We see a lot of 24-byte objects, which we know are tokens.

TODO: without filtering, uftrace should be able to tell us what other
objects are common, and which are 24 bytes.

devtools/services: Add asciinema.sh automation