[benchmarks refactor] Remove OIL_NATIVE variable
It was a failed abstraction attempt over provenance.

The whole 'filter-provenance' pattern has to go -- it is used in most
benchmarks.

Instead we should use the pattern that benchmarks/gc.sh does:

- Every shell bin path is unique: _bin/osh, bin/osh, etc.
- Then there can be a separate table with (shell bin path, hash), to
  join with provenance details
- Then we don't need to use xargs in print-tasks to retain provenance.
  We can simply use nested loops to print tasks, and then join later on
  shell_bin.

We also don't need to use xargs to execute tasks.  Instead we use a
'while read' loop.  This removes the possibility of parallel execution,
but it's OK for benchmarks.  We have parallelism at a higher level.