[mycpp] Implement object ID recycling, via lazy free()
This prevents the crash after 2^24 = 16 Mi object IDs.  I ran fib.sh for
5 seconds, doing 41 M allocations.

This should reduce GC pauses, hiding some latency in the mutator.

It might do more work overall, but it's a starting point, and we're
still way better off with the mark bitmap.

We might want to introduce our own pool allocator to get rid of the
common free() then malloc() case in Allocate().