[mycpp] emit code for generators
This commit is a first attempt at translating generator functions. If
mycpp sees a function that returns `typing.Iterator[T]` it will generate
a C++ function with a type signature that differs in the following ways:

  1) it returns `void` instead of `Iterator[T]`
  2) it has an extra ouput param of type `List<T>*`

The defintion of the function emitted by mycpp will be identical to its
pythonic ancestor, except all `yield ...` statements will be translated
into calls to `append(...)` on the output param from (2).

At the call sites for the function mycpp will allocate a temporary
`List<T>` on the stack and pass it as an ouput param to the modified
function as an accumulator the generator's output, which will be
computed in one big batch. All subsequent statements that refer to the
generator will operate, unmodified, on an `ListIter<T>` that wraps the
accumulator.

This approach described by @andychu here:
https://oilshell.zulipchat.com/#narrow/stream/121539-oil-dev/topic/Solutions.20to.20yield.20problem/near/316873052